update function view infor, view rental, rental-details
