+++ GreenMobility/Areas/Admin/Views/AdminRentals/Edit.cshtml
@@ -58,7 +58,7 @@
                 <div class="row">
                     <div class="col-md-12">
                         <div class="form-group">
-                            <label>Chọn trạng thái đơn hàng</label>
+                            <label class="font-weight-semibold" for="RentalStatusId">Chọn trạng thái đơn hàng</label>
                             <select class="form-control" asp-for="RentalStatusId" asp-items="@ViewBag.Status" required>
                                 <option value="0" disabled selected>Chọn trạng thái đơn hàng</option>
                             </select>
