--- GreenMobility/Controllers/AccountsController.cs
@@ -1,254 +1,0 @@
-﻿using AspNetCoreHero.ToastNotification.Abstractions;
-using AspNetCoreHero.ToastNotification.Notyf;
-using GreenMobility.Helper;
-using GreenMobility.Models;
-using GreenMobility.ViewModels;
-using Microsoft.AspNetCore.Authentication;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
-using System.Security.Claims;
-using GreenMobility.Extension;
-
-namespace GreenMobility.Controllers
-{
-    public class AccountsController : Controller
-    {
-        private readonly GreenMobilityContext _context;
-        public INotyfService _notyf { get; }
-        public AccountsController(GreenMobilityContext context, INotyfService notyf)
-        {
-            _context = context;
-            _notyf = notyf;
-        }
-        [HttpGet]
-        [AllowAnonymous]
-        public IActionResult ValidatePhone(string Phone)
-        {
-            try
-            {
-                var khachhang = _context.Customers.AsNoTracking().SingleOrDefault(x => x.Phone.ToLower() == Phone.ToLower());
-                if (khachhang != null)
-                    return Json(data: "Số điện thoại : " + Phone + "đã được sử dụng");
-
-                return Json(data: true);
-
-            }
-            catch
-            {
-                return Json(data: true);
-            }
-        }
-        [HttpGet]
-        [AllowAnonymous]
-        public IActionResult ValidateEmail(string Email)
-        {
-            try
-            {
-                var khachhang = _context.Customers.AsNoTracking().SingleOrDefault(x => x.Email.ToLower() == Email.ToLower());
-                if (khachhang != null)
-                    return Json(data: "Email : " + Email + " đã được sử dụng");
-                return Json(data: true);
-            }
-            catch
-            {
-                return Json(data: true);
-            }
-        }
-        [Route("tai-khoan-cua-toi.html", Name = "Dashboard")]
-        public IActionResult Dashboard()
-        {
-            var taikhoanID = HttpContext.Session.GetString("CustomerId");
-            if (taikhoanID != null)
-            {
-                var khachhang = _context.Customers.AsNoTracking().SingleOrDefault(x => x.CustomerId == Convert.ToInt32(taikhoanID));
-                if (khachhang != null)
-                {
-                    var lsDonHang = _context.Rentals
-                        .Include(x => x.RentalStatus)
-                        .AsNoTracking()
-                        .Where(x => x.CustomerId == khachhang.CustomerId)
-                        .ToList();
-                    ViewBag.DonHang = lsDonHang;
-                    return View(khachhang);
-                }
-
-            }
-            return RedirectToAction("Login");
-        }
-        [HttpGet]
-        [AllowAnonymous]
-        [Route("dang-ky.html", Name = "DangKy")]
-        public IActionResult Register()
-        {
-            return View();
-        }
-
-        [HttpPost]
-        [AllowAnonymous]
-        [Route("dang-ky.html", Name = "DangKy")]
-        public async Task<IActionResult> Register(RegisterVM taikhoan)
-        {
-            try
-            {
-                if (ModelState.IsValid)
-                {
-                    string salt = Utilities.GetRandomKey();
-                    Customer khachhang = new Customer
-                    {
-                        FullName = taikhoan.FullName,
-                        Phone = taikhoan.Phone.Trim().ToLower(),
-                        Email = taikhoan.Email.Trim().ToLower(),
-                        Password = (taikhoan.Password + salt.Trim()).ToMD5(),
-                        IsLocked = false,
-                        Salt = salt,
-                        CreateDate = DateTime.Now
-                    };
-                    try
-                    {
-                        _context.Add(khachhang);
-                        await _context.SaveChangesAsync();
-                        //Lưu Session MaKh
-                        HttpContext.Session.SetString("CustomerId", khachhang.CustomerId.ToString());
-                        var taikhoanID = HttpContext.Session.GetString("CustomerId");
-
-                        //Identity
-                        var claims = new List<Claim>
-                        {
-                            new Claim(ClaimTypes.Name,khachhang.FullName),
-                            new Claim("CustomerId", khachhang.CustomerId.ToString())
-                        };
-                        ClaimsIdentity claimsIdentity = new ClaimsIdentity(claims, "login");
-                        ClaimsPrincipal claimsPrincipal = new ClaimsPrincipal(claimsIdentity);
-                        await HttpContext.SignInAsync(claimsPrincipal);
-                        _notyf.Success("Đăng ký thành công");
-                        return RedirectToAction("Dashboard", "Accounts");
-                    }
-                    catch
-                    {
-                        return RedirectToAction("Register", "Accounts");
-                    }
-                }
-                else
-                {
-                    return View(taikhoan);
-                }
-            }
-            catch
-            {
-                return View(taikhoan);
-            }
-        }
-        [AllowAnonymous]
-        [Route("dang-nhap.html", Name = "DangNhap")]
-        public IActionResult Login(string returnUrl = null)
-        {
-            var taikhoanID = HttpContext.Session.GetString("CustomerId");
-            if (taikhoanID != null)
-            {
-                return RedirectToAction("Dashboard", "Accounts");
-            }
-            return View();
-        }
-        [HttpPost]
-        [AllowAnonymous]
-        [Route("dang-nhap.html", Name = "DangNhap")]
-        public async Task<IActionResult> Login(LoginVM customer, string returnUrl)
-        {
-            try
-            {
-                bool isEmail = Utilities.IsValidEmail(customer.UserName);
-                if (!isEmail) return View(customer);
-
-                var khachhang = _context.Customers.AsNoTracking().SingleOrDefault(x => x.Email.Trim() == customer.UserName);
-
-                if (khachhang == null) return RedirectToAction("Register");
-                string pass = (customer.Password + khachhang.Salt.Trim()).ToMD5();
-                if (khachhang.Password != pass)
-                {
-                    _notyf.Success("Thông tin đăng nhập chưa chính xác");
-                    return View(customer);
-                }
-                //kiem tra xem account co bi disable hay khong
-
-                if (khachhang.IsLocked == true)
-                {
-                    return RedirectToAction("ThongBao", "Accounts");
-                }
-
-                //Luu Session MaKh
-                HttpContext.Session.SetString("CustomerId", khachhang.CustomerId.ToString());
-                var taikhoanID = HttpContext.Session.GetString("CustomerId");
-
-                //Identity
-                var claims = new List<Claim>
-                    {
-                        new Claim(ClaimTypes.Name, khachhang.FullName),
-                        new Claim("CustomerId", khachhang.CustomerId.ToString())
-                    };
-                ClaimsIdentity claimsIdentity = new ClaimsIdentity(claims, "login");
-                ClaimsPrincipal claimsPrincipal = new ClaimsPrincipal(claimsIdentity);
-                await HttpContext.SignInAsync(claimsPrincipal);
-                _notyf.Success("Đăng nhập thành công");
-
-                if (string.IsNullOrEmpty(returnUrl))
-                {
-                    return RedirectToAction("Index", "Home");
-                }
-                else
-                {
-                    return Redirect(returnUrl);
-                }
-            }
-            catch
-            {
-                return RedirectToAction("Register", "Accounts");
-            }
-        }
-        [HttpGet]
-        [Route("dang-xuat.html", Name = "DangXuat")]
-        public IActionResult Logout()
-        {
-            HttpContext.SignOutAsync();
-            HttpContext.Session.Remove("CustomerId");
-            return RedirectToAction("Index", "Home");
-        }
-
-        [HttpPost]
-        public IActionResult ChangePassword(ChangePasswordVM model)
-        {
-            try
-            {
-                var taikhoanID = HttpContext.Session.GetString("CustomerId");
-                if (taikhoanID == null)
-                {
-                    return RedirectToAction("Login", "Accounts");
-                }
-                if (ModelState.IsValid)
-                {
-                    var taikhoan = _context.Customers.Find(Convert.ToInt32(taikhoanID));
-                    if (taikhoan == null) return RedirectToAction("Login", "Accounts");
-                    var pass = (model.PasswordNow.Trim() + taikhoan.Salt.Trim()).ToMD5();
-                    {
-                        string passnew = (model.Password.Trim() + taikhoan.Salt.Trim()).ToMD5();
-                        taikhoan.Password = passnew;
-                        _context.Update(taikhoan);
-                        _context.SaveChanges();
-                        _notyf.Success("Đổi mật khẩu thành công");
-                        return RedirectToAction("Dashboard", "Accounts");
-                    }
-                }
-            }
-            catch
-            {
-                _notyf.Success("Thay đổi mật khẩu không thành công");
-                return RedirectToAction("Dashboard", "Accounts");
-            }
-            _notyf.Success("Thay đổi mật khẩu không thành công");
-            return RedirectToAction("Dashboard", "Accounts");
-        }
-    }
-}
-
-
-
