+++ GreenMobility/Views/Account/Profile.cshtml
@@ -1,0 +1,64 @@
+﻿@model GreenMobility.Models.Customer
+
+@{
+    ViewData["Title"] = @_localization.Getkey("AccountInfor");
+    Layout = "~/Views/Shared/_Layout.cshtml";
+}
+<main class="main-content">
+    <div class="breadcrumb-area breadcrumb-height" data-bg-image="/assets/images/breadcrumb/bg/becdanhphoto.png">
+        <div class="container h-100">
+            <div class="row h-100">
+                <div class="col-lg-12">
+                    <div class="breadcrumb-item">
+                        <h2 class="breadcrumb-heading">@_localization.Getkey("MyAccount")</h2>
+                        <ul>
+                            <li>
+                                <a href="/">@_localization.Getkey("Home")<i class="pe-7s-angle-right"></i></a>
+                            </li>
+                            <li>@_localization.Getkey("MyAccount")</li>
+                        </ul>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+    <div class="account-page-area section-space-y-axis-100">
+        <div class="container">
+            <div class="row">
+                <div class="col-lg-3">
+                    @await Html.PartialAsync("_SideNavPartialView.cshtml")
+                </div>
+                <div class="col-lg-9">
+                    <div class="tab-pane fade show active" id="account-dashboard" role="tabpanel" aria-labelledby="account-dashboard-tab">
+                        <div class="myaccount-details">
+                            <form method="post" asp-controller="Account" asp-action="Profile" class="myaccount-form">
+                                <div class="myaccount-form-inner">
+                                    <div class="single-input">
+                                        <label for="FullName">@_localization.Getkey("FullName")</label>
+                                        <input type="text" class="form-control" id="FullName" name="FullName" value="@Model.FullName" required>
+                                        @Html.ValidationMessage("FullName", "", new { @style = "color:#f00" })
+                                    </div>
+                                    <div class="single-input single-input-half">
+                                        <label for="Email">@_localization.Getkey("EmailAddress")</label>
+                                        <input type="email" class="form-control" id="Email" name="Email" value="@Model.Email" required>
+                                        @Html.ValidationMessage("Email", "", new { @style = "color:#f00" })
+                                    </div>
+                                    <div class="single-input single-input-half">
+                                        <label for="Phone">@_localization.Getkey("Phone")</label>
+                                        <input type="text" class="form-control" id="Phone" name="Phone" value="@Model.Phone" required>
+                                        @Html.ValidationMessage("Phone", "", new { @style = "color:#f00" })
+                                    </div>
+                                    <div class="single-input">
+                                        <button class="btn btn-custom-size lg-size btn-secondary btn-primary-hover rounded-0" type="submit">
+                                            <span>@_localization.Getkey("Save")</span>
+                                        </button>
+                                    </div>
+                                </div>
+                            </form>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+</main>
