+++ GreenMobility/Views/Account/RentalList.cshtml
@@ -1,0 +1,90 @@
+﻿@model List<GreenMobility.Models.Rental>
+
+@{
+    ViewData["Title"] = @_localization.Getkey("AccountInfor");
+    Layout = "~/Views/Shared/_Layout.cshtml";
+}
+<main class="main-content">
+    <div class="breadcrumb-area breadcrumb-height" data-bg-image="/assets/images/breadcrumb/bg/becdanhphoto.png">
+        <div class="container h-100">
+            <div class="row h-100">
+                <div class="col-lg-12">
+                    <div class="breadcrumb-item">
+                        <h2 class="breadcrumb-heading">@_localization.Getkey("MyAccount")</h2>
+                        <ul>
+                            <li>
+                                <a href="/">@_localization.Getkey("Home")<i class="pe-7s-angle-right"></i></a>
+                            </li>
+                            <li>@_localization.Getkey("MyAccount")</li>
+                        </ul>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+    <div class="account-page-area section-space-y-axis-100">
+        <div class="container">
+            <div class="row">
+                <div class="col-lg-3">
+                    @await Html.PartialAsync("_SideNavPartialView.cshtml")
+                </div>
+                <div class="col-lg-9">
+                    <div class="myaccount-orders">
+                        <h4 class="small-title">@_localization.Getkey("RentalList")</h4>
+                        <div class="table-responsive">
+                            <table class="table table-bordered table-hover">
+                                <tbody>
+                                    <tr>
+                                        <th>Id</th>
+                                        <th>@_localization.Getkey("RentalDate")</th>
+                                        <th>@_localization.Getkey("Status")</th>
+                                        <th>@_localization.Getkey("TotalMoney")</th>
+                                        <th></th>
+                                    </tr>
+                                    @foreach (var item in Model)
+                                    {
+                                        <tr>
+                                            <td><a class="account-order-id">@item.RentalId </a></td>
+                                            <td>@item.OrderTime</td>
+                                            <td>@item.RentalStatus.Description</td>
+                                            <td>@item.TotalMoney?.ToString("#,##0") VNĐ</td>
+                                            <td>
+                                                <a href="javascript:void(0)" class="viewrental" data-rentalid="@item.RentalId">@_localization.Getkey("View")</a>
+                                            </td>
+                                        </tr>
+                                    }
+                                </tbody>
+                            </table>
+                        </div>
+                    </div>
+                    <hr />
+                    <div id="records_table">
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+</main>
+@section Scripts {
+    <script>
+        $(document).ready(function () {
+            $(".viewrental").click(function () {
+                var rentalid = $(this).attr("data-rentalid")
+                $.ajax({
+                    url: '/RentalDetails/Details',
+                    datatype: "json",
+                    type: "POST",
+                    data: { id: rentalid },
+                    async: true,
+                    success: function (results) {
+                        $("#records_table").html("");
+                        $("#records_table").html(results);
+                    },
+                    error: function (xhr) {
+                        console.log(xhr);
+                    }
+                });
+            });
+        });
+    </script>
+}
