+++ GreenMobility/Views/Account/_SideNavPartialView.cshtml
@@ -1,0 +1,14 @@
+﻿<ul class="nav myaccount-tab-trigger" id="account-page-tab" role="tablist">
+    <li class="nav-item">
+        <a class="nav-link" asp-controller="Account" asp-action="Profile">@_localization.Getkey("AccountInfor")</a>
+    </li>
+    <li class="nav-item">
+        <a class="nav-link" asp-controller="Account" asp-action="RentalList">@_localization.Getkey("RentalList")</a>
+    </li>
+    <li class="nav-item">
+        <a class="nav-link" href="#account-details" role="tab">@_localization.Getkey("ChangePassword")</a>
+    </li>
+    <li class="nav-item">
+        <a class="nav-link" href="/logout.html" role="tab" aria-selected="false">@_localization.Getkey("Logout")</a>
+    </li>
+</ul>
