--- GreenMobility/Views/Accounts/Dashboard.cshtml
@@ -1,126 +1,0 @@
-﻿@model GreenMobility.Models.Customer
-
-@{
-    ViewData["Title"] = "Dashboard";
-    Layout = "~/Views/Shared/_Layout.cshtml";
-    List<Rental> DanhSachDonHang = ViewBag.DonHang;
-
-}
-@* GreenMobility.ViewModels.ChangePasswordViewModel changePassword = new GreenMobility.ViewModels.ChangePasswordViewModel(); *@
-<main class="main-content">
-    <div class="breadcrumb-area breadcrumb-height" data-bg-image="assets/images/breadcrumb/bg/1-1-1920x373.jpg">
-        <div class="container h-100">
-            <div class="row h-100">
-                <div class="col-lg-12">
-                    <div class="breadcrumb-item">
-                        <h2 class="breadcrumb-heading">@_localization.Getkey("MyAccount")</h2>
-                        <ul>
-                            <li>
-                                <a href="/">@_localization.Getkey("Home")<i class="pe-7s-angle-right"></i></a>
-                            </li>
-                            <li>@_localization.Getkey("MyAccount")</li>
-                        </ul>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-    <div class="account-page-area section-space-y-axis-100">
-        <div class="container">
-            <div class="row">
-                <div class="col-lg-3">
-                    <ul class="nav myaccount-tab-trigger" id="account-page-tab" role="tablist">
-                        <li class="nav-item">
-                            <a class="nav-link active" id="account-dashboard-tab" data-bs-toggle="tab" href="#account-dashboard" role="tab" aria-controls="account-dashboard" aria-selected="true">@_localization.Getkey("AccountInfor")</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link" id="account-orders-tab" data-bs-toggle="tab" href="#account-orders" role="tab" aria-controls="account-orders" aria-selected="false">@_localization.Getkey("RentalList")</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link" id="account-details-tab" data-bs-toggle="tab" href="#account-details" role="tab" aria-controls="account-details" aria-selected="false">@_localization.Getkey("ChangePassword")</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link" id="account-logout-tab" href="/dang-xuat.html" role="tab" aria-selected="false">@_localization.Getkey("Home")</a>
-                        </li>
-                    </ul>
-                </div>
-                <div class="col-lg-9">
-                    <div class="tab-content myaccount-tab-content" id="account-page-tab-content">
-                        <div class="tab-pane fade show active" id="account-dashboard" role="tabpanel" aria-labelledby="account-dashboard-tab">
-                            <div class="myaccount-dashboard">
-                                <p>
-                                    @_localization.Getkey("Hi") <b>@Model.FullName</b>
-                                </p>
-                                <p>
-                                    @_localization.Getkey("EmailAddress"): @Model.Email
-                                </p>
-                                <p>
-                                    @_localization.Getkey("Phone"): @Model.Phone
-                                </p>
-                            </div>
-                        </div>
-                        <div class="tab-pane fade" id="account-orders" role="tabpanel" aria-labelledby="account-orders-tab">
-                            <div class="myaccount-orders">
-                                <h4 class="small-title">@_localization.Getkey("RentalList")</h4>
-                                <div class="table-responsive">
-                                    <table class="table table-bordered table-hover">
-                                        <tbody>
-                                            <tr>
-                                                <th>Id</th>
-                                                <th>@_localization.Getkey("RentalDate")</th>
-                                                <th>@_localization.Getkey("Status")</th>
-                                                <th>@_localization.Getkey("TotalMoney")</th>
-                                                <th></th>
-                                            </tr>
-                                            @foreach (var item in DanhSachDonHang)
-                                            {
-                                                <tr>
-                                                    <td><a class="account-order-id">@item.RentalId </a></td>
-                                                    <td>@item.OrderTime</td>
-                                                    <td>@item.RentalStatus.Description</td>
-                                                    <td>@item.TotalMoney?.ToString("#,##0") VNĐ</td>
-                                                    <td>
-                                                        <a href="javascript:void(0)" class="xemdonhang" data-madonhang="@item.RentalId">@_localization.Getkey("View")</a>
-                                                    </td>
-                                                </tr>
-                                            }
-                                        </tbody>
-                                    </table>
-                                </div>
-                            </div>
-                            <hr />
-                            <div id="records_table">
-                            </div>
-                        </div>
-                        <div class="tab-pane fade" id="account-details" role="tabpanel" aria-labelledby="account-details-tab">
-                            @*                             @await Html.PartialAsync("_ChangePasswordPartialView", changePassword) *@
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-</main>
-@section Scripts {
-    <script>
-        $(document).ready(function () {
-            $(".xemdonhang").click(function () {
-                var madonhang = $(this).attr("data-madonhang")
-                $.ajax({
-                    url: '/RentalDetails/Details',
-                    datatype: "json",
-                    type: "POST",
-                    data: { id: madonhang },
-                    async: true,
-                    success: function (results) {
-                        $("#records_table").html("");
-                        $("#records_table").html(results);
-                    },
-                    error: function (xhr) {
-                        console.log(xhr);
-                    }
-                });
-            });
-        });
-    </script>
-}
