+++ GreenMobility/Views/RentalCart/Index.cshtml
@@ -120,7 +120,7 @@
                                             }
                                             else
                                             {
-                                                <a href="/dang-nhap.html?returnUrl=/checkout.html" class="btn btn-secondary btn-primary-hover">Đến Trang Thanh toán</a>
+                                                <a href="/login.html?returnUrl=/checkout.html" class="btn btn-secondary btn-primary-hover">Đến Trang Thanh toán</a>
                                             }
                                         }
                                     </div>
