+++ GreenMobility/Views/Rentals/List.cshtml
@@ -7,7 +7,7 @@
     int PageNext = PageCurrent + 1;
 }
 <main class="main-content">
-    <div class="breadcrumb-area breadcrumb-height" data-bg-image="assets/images/breadcrumb/bg/becdanhphoto.png">
+    <div class="breadcrumb-area breadcrumb-height" data-bg-image="/assets/images/breadcrumb/bg/becdanhphoto.png">
         <div class="container h-100">
             <div class="row h-100">
                 <div class="col-lg-12">
