+++ GreenMobility/Views/Shared/Components/NumberCart/Default.cshtml
@@ -2,7 +2,7 @@
 @if (Model != null && Model.Count() > 0)
 {
     <li class="minicart-wrap me-3 me-lg-0">
-        <a href="cart.html">
+        <a href="/cart.html">
             <i class="pe-7s-shopbag"></i>
             <span class="quantity">@Model.Count()</span>
         </a>
@@ -11,7 +11,7 @@
 else
 {
     <li class="minicart-wrap me-3 me-lg-0">
-        <a href="cart.html">
+        <a href="/cart.html">
             <i class="pe-7s-shopbag"></i>
         </a>
     </li>
